Update Python deps (#6646)
* Add --no-plugins switch to dep compile tool

Especially when testing specific package updates, it's annoying and
slower to get changes in plugins at the same time

* Update Python deps

- removed the usage of OrderedMultiDict; this should be fine since
  dicts are ordered anyway nowadays, and even the weird case in the
  paypal plugin (where the args have to sent back for verification in
  the same order) should work as expected because of this

+++ indico/modules/rb/models/reservation_occurrences.py
@@ -220,7 +220,7 @@ class ReservationOccurrence(db.Model):
         elif repeat_frequency == RepeatFrequency.MONTH:
             if repeat_interval == 0:
                 raise IndicoError('Unsupported interval')
-            position = int(ceil(start.day / 7.0))
+            position = ceil(start.day / 7.0)
             if position == 5:
                 # The fifth weekday of the month will always be the last one
                 position = -1
